some refactoring and tidying
