+++ src/nkp_cluster_cleaner/notify_command.py
@@ -3,118 +3,13 @@ Notify command implementation for the NKP Cluster Cleaner tool.
 """
 
 import click
-import redis
 from datetime import datetime, timedelta
 from colorama import Fore, Style
 from tabulate import tabulate
 from typing import Optional, List, Tuple, Dict, Any, Set
 from .config import ConfigManager
 from .notification_manager import NotificationManager
-
-
-# Supported notification backends
-SUPPORTED_BACKENDS = ["slack"]
-
-
-class NotificationHistory:
-    """Manages notification history using Redis to prevent duplicate alerts."""
-    
-    def __init__(self, redis_host: str = 'redis', redis_port: int = 6379, redis_db: int = 0):
-        """
-        Initialize notification history manager.
-        
-        Args:
-            redis_host: Redis host
-            redis_port: Redis port  
-            redis_db: Redis database number
-        """
-        self.redis_client = redis.Redis(
-            host=redis_host,
-            port=redis_port,
-            db=redis_db,
-            decode_responses=True,
-            socket_connect_timeout=5,
-            socket_timeout=5,
-            retry_on_timeout=True
-        )
-        
-        # Test connection
-        try:
-            self.redis_client.ping()
-        except redis.ConnectionError as e:
-            raise Exception(f"Failed to connect to Redis at {redis_host}:{redis_port}: {e}")
-    
-    def _get_cluster_key(self, cluster_name: str, namespace: str) -> str:
-        """Generate Redis key for cluster notification history."""
-        return f"notifications:cluster:{namespace}:{cluster_name}"
-    
-    def has_been_notified(self, cluster_name: str, namespace: str, severity: str) -> bool:
-        """
-        Check if a cluster has already been notified at this severity level.
-        
-        Args:
-            cluster_name: Name of the cluster
-            namespace: Namespace of the cluster
-            severity: "warning" or "critical"
-            
-        Returns:
-            True if already notified at this severity level
-        """
-        key = self._get_cluster_key(cluster_name, namespace)
-        return self.redis_client.sismember(key, severity)
-    
-    def mark_as_notified(self, cluster_name: str, namespace: str, severity: str, ttl_days: int = 30):
-        """
-        Mark a cluster as notified at the given severity level.
-        
-        Args:
-            cluster_name: Name of the cluster
-            namespace: Namespace of the cluster
-            severity: "warning" or "critical"
-            ttl_days: Number of days to keep the notification record
-        """
-        key = self._get_cluster_key(cluster_name, namespace)
-        
-        # Add severity to the set for this cluster
-        self.redis_client.sadd(key, severity)
-        
-        # Set TTL for the key
-        self.redis_client.expire(key, ttl_days * 24 * 3600)
-    
-    def filter_new_notifications(self, clusters: List[Tuple], severity: str) -> List[Tuple]:
-        """
-        Filter out clusters that have already been notified at this severity level.
-        
-        Args:
-            clusters: List of (cluster_info, elapsed_percentage, expiry_time) tuples
-            severity: "warning" or "critical"
-            
-        Returns:
-            List of clusters that haven't been notified yet
-        """
-        new_clusters = []
-        
-        for cluster_info, elapsed_percentage, expiry_time in clusters:
-            cluster_name = cluster_info.get("capi_cluster_name", "unknown")
-            namespace = cluster_info.get("capi_cluster_namespace", "unknown")
-            
-            if not self.has_been_notified(cluster_name, namespace, severity):
-                new_clusters.append((cluster_info, elapsed_percentage, expiry_time))
-        
-        return new_clusters
-    
-    def mark_clusters_as_notified(self, clusters: List[Tuple], severity: str):
-        """
-        Mark multiple clusters as notified.
-        
-        Args:
-            clusters: List of (cluster_info, elapsed_percentage, expiry_time) tuples
-            severity: "warning" or "critical"
-        """
-        for cluster_info, elapsed_percentage, expiry_time in clusters:
-            cluster_name = cluster_info.get("capi_cluster_name", "unknown")
-            namespace = cluster_info.get("capi_cluster_namespace", "unknown")
-            self.mark_as_notified(cluster_name, namespace, severity)
+from .notification_history import NotificationHistory
 
 
 def _send_notifications(critical_clusters: List[Tuple], warning_clusters: List[Tuple], 
@@ -232,8 +127,8 @@ def execute_notify_command(kubeconfig: Optional[str], config: Optional[str], nam
         **kwargs: Backend-specific parameters (e.g. slack_token, slack_channel for slack backend)
     """
     # Validate notification backend
-    if notify_backend and notify_backend not in SUPPORTED_BACKENDS:
-        click.echo(f"{Fore.RED}Error: Unsupported notification backend '{notify_backend}'. Supported backends: {', '.join(SUPPORTED_BACKENDS)}{Style.RESET_ALL}")
+    if notify_backend and notify_backend not in NotificationManager.SUPPORTED_BACKENDS:
+        click.echo(f"{Fore.RED}Error: Unsupported notification backend '{notify_backend}'. Supported backends: {', '.join(NotificationManager.SUPPORTED_BACKENDS)}{Style.RESET_ALL}")
         raise click.Abort()
     
     # Validate backend-specific requirements
